feat(docs): replace hand-built Algolia search with published plugin
The docs site shipped a hand-built Algolia search implemented directly in
the Docusaurus theme. That same search is now maintained and published as
@mastra/docusaurus-plugin-algolia, so consume the package instead of
carrying a local copy that drifts from upstream.

Register the plugin in docusaurus.config.ts (preserving the docs_v1_crawler
index and porting the empty-state links into suggestedLinks), wire its
@theme/SearchBar into the navbar, and keep the Ask AI button by extracting
it into its own component. The plugin's design tokens are provided by the
navbar/sidebar styles branch this builds on; only the theme helper classes
(.mastra-search-button, .mastra-card, .dialog-panel) it expects are mapped
onto the docs' own tokens. Delete the now-unused hand-built search files and
the direct algoliasearch/dompurify dependencies.
