+++ docs/package.json
@@ -48,6 +48,7 @@
     "@headlessui/react": "^2.2.9",
     "@hookform/resolvers": "^5.2.2",
     "@kapaai/react-sdk": "^0.9.2",
+    "@mastra/docusaurus-plugin-algolia": "^1.1.1",
     "@mdx-js/react": "^3.1.1",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -57,10 +58,8 @@
     "@radix-ui/react-slot": "^1.2.4",
     "@tanstack/react-virtual": "^3.13.22",
     "@vercel/analytics": "^2.0.1",
-    "algoliasearch": "^5.49.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "dompurify": "^3.4.11",
     "dotenv": "^17.3.1",
     "fs-extra": "^11.3.4",
     "hast-util-select": "^6.0.4",
