--- docs/src/theme/Navbar/Search/styles.module.css
@@ -1,20 +1,0 @@
-/*
-Workaround to avoid rendering empty search container
-See https://github.com/facebook/docusaurus/pull/9385
-*/
-.navbarSearchContainer:empty {
-  display: none;
-}
-
-@media (max-width: 62.25rem) {
-  .navbarSearchContainer {
-    position: absolute;
-    right: var(--ifm-navbar-padding-horizontal);
-  }
-}
-
-@media (min-width: 62.25rem) {
-  .navbarSearchContainer {
-    padding: 0 var(--ifm-navbar-item-padding-horizontal);
-  }
-}
