+++ docs/src/theme/Navbar/index.tsx
@@ -2,8 +2,9 @@ import Link from '@docusaurus/Link'
 import { GitHubIconLink } from '@site/src/components/github-icon-link'
 import NavbarLayout from '@theme/Navbar/Layout'
 import NavbarMobileSidebarToggle from '@theme/Navbar/MobileSidebar/Toggle'
+import SearchBar from '@theme/SearchBar'
 import { type ReactNode } from 'react'
-import SearchContainer, { AskAI } from './Search'
+import { AskAI } from './ask-ai'
 import { Logo } from './logo'
 import { TabSwitcher } from './tab-switcher'
 
@@ -25,7 +26,7 @@ function NavbarContentDesktop() {
         </div>
 
         <div className="flex items-center gap-2">
-          <SearchContainer locale="en" />
+          <SearchBar />
           <div className="hidden lg:block">
             <AskAI />
           </div>
